Added the risk analysis service
Added this service that will read settled bets, unsettled bets, generate
client statistics and search for settled bets by id.
A new interface was added -IFilePathsProvider- that has to be
implemented by the client of the domain library. It specifies where are
the files for settled and unsettled. Currently we expect only 1 file but
can be easily extended to support more files of each type. Also added
tests.

+++ RiskApplication/Domain/Models/CustomerStatics.cs
@@ -3,7 +3,7 @@
     public class CustomerStatics
     {
         public int Customer { get; set; }
-        public decimal AverageStake { get; set; }
+        public double AverageStake { get; set; }
         public int TotalBets { get; set; }
         public int TotalBetsWon { get; set; }
